feat:add New password system, Trainings and exercises

+++ migrations/m260305_113103_create_exercises_table.php
@@ -12,18 +12,19 @@ class m260305_113103_create_exercises_table extends Migration
      */
     public function safeUp()
     {
-       $this->createTable('{{%exercises}}', [
-            'id' => $this->primaryKey(),
-            'name' => $this->string(255)->notNull(),
-            'type' => $this->string(50)->notNull(), // 
-            'sets' => $this->integer()->defaultValue(null),
-            'reps' => $this->integer()->defaultValue(null),
-            'duration_sec' => $this->integer()->defaultValue(null), 
-            'rest_sec' => $this->integer()->defaultValue(60),
-            'created_at' => $this->integer()->notNull(),
-            'updated_at' => $this->integer()->notNull(),
-        ]);
-      
+        if ($this->db->schema->getTableSchema('{{%exercises}}', true) === null) {
+            $this->createTable('{{%exercises}}', [
+                'id' => $this->primaryKey(),
+                'name' => $this->string(255)->notNull(),
+                'type' => $this->string(50)->notNull(), // 
+                'sets' => $this->integer()->defaultValue(null),
+                'reps' => $this->integer()->defaultValue(null),
+                'duration_sec' => $this->integer()->defaultValue(null), 
+                'rest_sec' => $this->integer()->defaultValue(60),
+                'created_at' => $this->integer()->notNull(),
+                'updated_at' => $this->integer()->notNull(),
+            ]);
+        }
     }
 
     /**
@@ -31,6 +32,8 @@ class m260305_113103_create_exercises_table extends Migration
      */
     public function safeDown()
     {
-         $this->dropTable('{{%exercises}}');
+        if ($this->db->schema->getTableSchema('{{%exercises}}', true) !== null) {
+            $this->dropTable('{{%exercises}}');
+        }
     }
 }
